env vars become a dictionary and values overrided
